chore: use member function instead of local lambda
The lambda did not save the changed index, while the member function did.
Instead of moving saving from the method to the lambda, use the existing
function. In this case, getting the parameter type for the member
function requires less code than having a dedicated function.

- use member function instead of local lambda
- change data type for index data member from int to long
    While lists longer than maximum integer are not expected for language
    resources, maybe following along with a usual index type is cleaner?

+++ include/tts/tts_language_settings.hpp
@@ -17,10 +17,10 @@ public:
     LanguageSettings();
 
     QStringList availableLanguages();
-    int index();
+    long index();
     bool isInAutoMode();
 
-    void setIndex(const int i);
+    void setIndex(const long i);
     void setToAutoMode();
     void setToManualMode();
 
@@ -31,7 +31,7 @@ private:
 
     bool isInAutoMode_;
     QStringList languages_;
-    int index_;
+    long index_;
 };
 
 } // namespace Tts
